core: add toml/target monitoring
This adds a thread to core that checks when reactor was reloaded the
last time and compares modified timestamps of toml and target files (for
the promoter plugin) accordingly. In case there is a mismatch a warning
is printed to the syslog (e.g., when the toml is newer than the last
reload).

This can be disabled by setting snippets-monitoring-interval to 0.

As we don't expect ten or hundred thousands of toml files stat-ing the
files should be fine. Currently an extra dependency on some inotify lib
did not seem to be worth it. We can revise that later if necessary.